make inc_decoding work
