+++ examples/cpp/inference/mixture_of_experts/moe.cc
@@ -15,6 +15,7 @@
 
 #include "moe.h"
 #include "flexflow/inference.h"
+#include "flexflow/request_manager.h"
 #include <cstdlib>
 #include <fstream>
 #include <iostream>
