+++ include/flexflow/batch_config.h
@@ -52,7 +52,7 @@ public:
   void print() const;
   virtual InferenceMode get_mode() const;
   static BatchConfig const *from_future(BatchConfigFuture const &future);
-  static int const MAX_NUM_REQUESTS = 16;
+  static int const MAX_NUM_REQUESTS = 1;
   static int const MAX_NUM_TOKENS = 64;
   static int const MAX_SEQ_LENGTH = 256;
