+++ inference/models/opt.h
@@ -17,6 +17,7 @@
 #include "file_loader.h"
 #include "flexflow/batch_config.h"
 #include "flexflow/inference.h"
+#include "flexflow/request_manager.h"
 #include <nlohmann/json.hpp>
 #include <string>
 using json = nlohmann::json;
