+++ src/mapper/mapper.cc
@@ -286,7 +286,8 @@ void FFMapper::select_task_options(const MapperContext ctx,
   if ((task.task_id == RM_PREPARE_NEXT_BATCH_TASK_ID) ||
       (task.task_id == RM_PREPARE_NEXT_BATCH_BEAM_TASK_ID) ||
       (task.task_id == RM_PREPARE_NEXT_BATCH_INIT_TASK_ID) ||
-      (task.task_id == RM_PREPARE_NEXT_BATCH_VERIFY_TASK_ID)) {
+      (task.task_id == RM_PREPARE_NEXT_BATCH_VERIFY_TASK_ID) ||
+      (task.task_id == RM_LLM_SERVING_BACKGROUND_TASK_ID)) {
     output.initial_proc = all_cpus[0];
     return;
   }
